+++ test/json_premium_index.cpp
@@ -30,5 +30,16 @@ TEST(json_premium_index, simple_swap) {
   core::Buffer buffer(8192);
   core::json::Buffer buffer_(buffer);
   auto obj = core::json::Parser::create<json::PremiumIndex>(message, buffer_);
+  EXPECT_EQ(obj.ch, "market.BTC-USD.premium_index.1min"sv);
+  EXPECT_EQ(obj.ts, 1642657680747ms);
+  auto &tick = obj.tick;
+  EXPECT_EQ(tick.id, 1642657680);
+  EXPECT_EQ(tick.id, 1642657680);
+  EXPECT_DOUBLE_EQ(tick.open, -0.0002871886442248);
+  EXPECT_DOUBLE_EQ(tick.close, -0.0002871886442248);
+  EXPECT_DOUBLE_EQ(tick.high, -0.0002871886442248);
+  EXPECT_DOUBLE_EQ(tick.low, -0.0002871886442248);
+  EXPECT_DOUBLE_EQ(tick.amount, 0.0);
+  EXPECT_DOUBLE_EQ(tick.vol, 0.0);
+  EXPECT_DOUBLE_EQ(tick.count, 0.0);
 }
-