+++ test/json_subbed.cpp
@@ -12,8 +12,8 @@ using namespace roq::huobi_futures;
 using namespace std::literals;
 using namespace std::chrono_literals;
 
-// note! reduced
 TEST(json_subbed, simple) {
+  /*
   auto message = R"({)"
                  R"("id":"3000001",)"
                  R"("subbed":"market.FIL211231.bbo",)"
@@ -21,4 +21,5 @@ TEST(json_subbed, simple) {
                  R"("status":"ok")"
                  R"(})";
   // auto obj = core::json::Parser::create<json::Subbed>(message, buffer_);
+  */
 }